requirements.txt
flask>=2.3.0
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=9.0.0
werkzeug>=2.3.0
gunicorn>=21.0.0
